requirements.txt
-r requirements-api.txt
-r requirements-dashboard.txt
-r requirements-ml.txt
pyserial==3.5
